refactor: move row document actions up to the actor sheet base
viewDoc / createDoc / deleteDoc が共鳴者シートとNPCシートに二重に実装されて
いた。createDoc は完全に同一で、viewDoc と deleteDoc は実装が違う — 共鳴者は
utils/sheet.ts の getEmbeddedDocument（アイテムと効果の両方を引ける）、NPCは
手書きの closest("[data-item-id]") を辿っていた。

3つとも EmokloreActorSheet へ引き上げ、getEmbeddedDocument 版に揃える。
基底は行の右クリックメニューで既に同じ関数を通っているので、行の解決が
1本になる。怪異シートはこれらの data-action を持たないので発火しない。
